% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api}
\alias{api}
\title{Get data from API.}
\usage{
api(url, path = NULL, parms = NULL, do.cache = TRUE, cache.path = NULL,
  cache.name = "api-cache", use.cache.same.day = TRUE, verbose = TRUE)
}
\arguments{
\item{url}{The URL of the API. You can set this up hou you like but it usually starts with htt and ends with .com or similar.}

\item{path}{The second part of the URL which varies by the service you are using.}

\item{do.cache}{Save a cache of the output to be used instead of the API in case of an error, etc.}

\item{cache.path}{Path to save a cache to. Leave this blank and just use cache name if you wnat.}

\item{cache.name}{File name to use for the cache. Cache will always be a .RDS so you don't need to include the path.}

\item{use.cache.same.day}{Use the cache instead of calling the API again if the cache is the same day. Useful when you have limited API calls to an API.}

\item{verbose}{Print helpful messages about the API.}

\item{params}{Named list/vector of key value pairs representing the arguments.}
}
\value{
List of data from the API.
}
\description{
Get data from API.
}
\examples{
#TODO
}
