% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arOrderSelection.R
\name{arOrderSelection}
\alias{arOrderSelection}
\title{Order selection in autoregressive modeling for time series}
\usage{
arOrderSelection(x, maxOrder = NULL, criteria = c("BC"),
  fitMethod = "yule-walker", penaltyBC = NULL)
}
\arguments{
\item{x}{Vector of 1d time series data}

\item{maxOrder}{Integer of the maximum  order under consideration}

\item{criteria}{Vector of strings ('AIC, BIC, BC') indicating criterion to use
in selecting the final subset}

\item{fitMethod}{String ("yule-walker", "burg", "ols", "mle", or "yw") indicating which method to use for fitting the autoregression.}

\item{penaltyBC}{Vector of non-default penalty values in using BC, default to NULL so that
only the suggested value n^(1/3) will be considered}
}
\value{
Vector (cri_opt) of indices of candidate models

Vector (cri_name) of the criteria used

Vector (PI) of parametricness index, default to one value that corresponds to 
the default BC penalty
}
\description{
For a sequence of 1d time series, 
this function can be used to fit autoregressive model and select the most appropriate order.
}
