% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behaviorchange.R
\docType{package}
\name{behaviorchange-package}
\alias{behaviorchange}
\alias{behaviorchange-package}
\title{behaviorchange: Tools for Behavior Change Researchers and Professionals}
\description{
Contains specialised analyses and
visualisation tools for behavior change science.
These facilitate conducting determinant studies
(for example, using confidence interval-based
estimation of relevance, CIBER, or CIBERlite
plots) and systematically developing, reporting,
and analysing interventions (for example, using
acyclic behavior change diagrams or ABCDs). This
package is especially useful for researchers in
the field of behavior change or health psychology
and to behavior change professionals such as
intervention developers and prevention workers.
}
\details{
Contains specialised analyses and
visualisation tools for behavior change science.
These facilitate conducting determinant studies
(for example, using confidence interval-based
estimation of relevance, CIBER, or CIBERlite
plots) and systematically developing, reporting,
and analysing interventions (for example, using
acyclic behavior change diagrams). This package
is especially useful for researchers in the field of
behavior change or health psychology and to
behavior change professionals such as intervention
developers and prevention workers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://a-bc.eu/R/behaviorchange}
  \item Report bugs at \url{https://github.com/academy-of-behavior-change/behaviorchange/issues}
}

}
\author{
\strong{Maintainer}: Gjalt-Jorn Peters \email{gjalt-jorn@a-bc.eu} [contributor]

}
\keyword{internal}
