% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockCreate.r
\name{blockCreate}
\alias{blockCreate}
\title{A function to create crossing blocks when the male is known}
\usage{
blockCreate(x, ge.male, nFemale, nMale)
}
\arguments{
\item{x, }{ge,male, nFamle, nMale}
}
\value{
a data frame
}
\description{
This function creates crossing block layout when the male parent is known.
}
\examples{
blockCreate(df, "Line001", 8, 4)

}
