% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat2bool.R
\name{cat2bool}
\alias{cat2bool}
\title{Convert Categorical Variables to Booleans.}
\usage{
cat2bool(x, match.to = NULL, make.names = TRUE, ignore.cols = c())
}
\arguments{
\item{x}{Dataframe or tibble.}

\item{match.to}{Data frame to match the values to. This will ensure consistent columns for modeling.}

\item{ignore.cols}{Columns to ignore.}
}
\value{
Modified data.
}
\description{
Convert Categorical Variables to Booleans.
}
\examples{
cat2bool( iris )
}
