% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.R
\name{cc}
\alias{cc}
\title{Easy Concatenate}
\usage{
cc(..., sep = "")
}
\arguments{
\item{...}{Arguments to be pasted. Eventually passed to paste.}

\item{sep}{Separator passed to paste.}
}
\value{
Concatenated result.
}
\description{
I got tired of typing paste and paste0 so created this function to replace them.
}
\examples{
cc( c( '1', '2', '3' ) ) 
}
