% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellBarcode.R
\docType{package}
\name{CellBarcode}
\alias{CellBarcode-package}
\alias{CellBarcode}
\title{DNA Barcode Analysis toolkit}
\description{
The package CellBarcode performs Cellular DNA Barcode analysis. It can handle all kinds of DNA barcodes, as long as the barcode is within a single sequencing read and has a pattern that can be matched by a regular expression. \code{CellBarcode} can handle barcodes with flexible lengths, with or without UMI (unique molecular identifier). This tool also can be used for pre-processing some amplicon data such as CRISPR gRNA screening, immune repertoire sequencing, and metagenome data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://wenjie1991.github.io/CellBarcode/}
  \item Report bugs at \url{https://github.com/wenjie1991/CellBarcode/issues}
}

}
\author{
\strong{Maintainer}: Wenjie Sun \email{sunwjie@gmail.com} (\href{https://orcid.org/0000-0002-3100-2346}{ORCID})

Authors:
\itemize{
  \item Anne-Marie Lyne \email{Anne-Marie.Lyne@curie.fr}
  \item Leila Perie \email{leila.perie@curie.fr}
}

}
