% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chuck.r
\name{chunk}
\alias{chunk}
\title{A function to divide a data frame (x) into equal row number (n) of data frames (by row)}
\usage{
chunk(x, n)
}
\arguments{
\item{x, }{n}
}
\value{
a list of data frames
}
\description{
A function to divide a data frame (x) into equal row number (n) of data frames (by row).
}
\examples{
chunk(df, 3)

}
