% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Clean Data}
\usage{
clean(x, header.in.data = TRUE, header.on.row = NULL)
}
\arguments{
\item{x}{Data frame to clean.}

\item{header.in.data}{Column names are in the data and must be extracted from it.}

\item{header.on.row}{Column names are on this row.}
}
\value{
Clean tibble.
}
\description{
My standard methods for cleaning a dataset.
}
\examples{
#TODO
}
