% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparePerformance.R
\name{comparePerformance}
\alias{comparePerformance}
\title{Compare MCP, SCAD, LASSO, and BCVO for high dimensional regression}
\usage{
comparePerformance(dat, dat_test, hd_methods, dfmax, ratio, weight_method,
  criteria = c("GIC2", "AIC", "BIC", "BC"), penaltyBC = NULL,
  adaptiveBC = TRUE, methodPI = "BC")
}
\arguments{
\item{dat}{Matrix of design that is used for training}

\item{dat_test}{Matrix of design that is used for comparing performance}

\item{hd_methods}{Vector of strings ("MCP", "SCAD", "lasso") indicating which penalized regression
methods are used to generate initial candidate subsets}

\item{dfmax}{Integer of maximum number of variables selected by each 
penalized regression}

\item{ratio}{Value (0 to 1) of proportion of data in obtaining initial 
candidate subsets}

\item{weight_method}{String indicating which method to use for weighting}

\item{criteria}{Vector of strings ('GIC2, GICn, Cp, AIC, BIC, BC') indicating criteria to use
in selecting the final subset}

\item{penaltyBC}{Vector of non-default penalty values in using BC, default to NULL so that
only the suggested value n^(1/3) will be considered}

\item{adaptiveBC}{Boolean indicating whether to use adaptive BC, default to FALSE so that
only the suggested value n^(1/3) will be considered}

\item{methodPI}{String ('BC', 'Drop', or 'Both') indicating the method to calculate BC, default to 'BC'}
}
\value{
List of prediction loss, prediction correlation, number of overfitted variables,  
number of underfitted variables, prediction residual, and selected subsets
}
\description{
This function compares the performance of MCP, SCAD, LASSO, and BCVO on
on a single dataset, in terms of prediction loss and over/under-fitting.
}
