% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc.R
\name{create_task}
\alias{create_task}
\title{Create Task (Define Data)}
\usage{
create_task(data, target, perc = 0.01, beta)
}
\arguments{
\item{data}{dataset used to train}

\item{target}{name of response variable}

\item{perc}{the percentage of a feature values in [0,1) that must differ from the mode value}

\item{beta}{response rate in original dataset (number of responses/number of non-responses)}
}
\description{
Create Task (Define Data)
}

