% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delayCal.r
\name{delayCal}
\alias{delayCal}
\title{A function to calculate delays within a crossing block}
\usage{
delayCal(datLayout, blk, nFemale, nMale, delay.start, delay.increment)
}
\arguments{
\item{df, }{block, nFemale, nMale, delay.start, delay.increament}
}
\value{
a data frame
}
\description{
A function to calculate delays within a crossing block.
}
\examples{
delayCal(Layout, 1, 8, 4, 50, 75)

}
