% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\name{dict}
\alias{dict}
\title{Data dictionary.}
\usage{
dict(x, nastrings = bc::nastrings, run.on.sample = NULL,
  do.droplevels = FALSE)
}
\arguments{
\item{x}{Data frame to summarize.}

\item{nastrings}{Strings to consider NA. Uses this package's defaults.}

\item{run.on.sample}{Optionally, sample data to make this run faster. Enter the % of rows to sample.}

\item{do.droplevels}{Unused factorlevels can cause problems with this. Choose to re-leveling all factors. This is slower but may resolve errors.}
}
\value{
Data frame with information for each column. 
Outliers use 1.5 IQR method. 
pareto80 is a measure of categorical concentration. It is the minimum % of categories that represent 80% of the rows. 80/20 rule applies if this is 0.2 for example. If it is 0.1, then 10% of categories make up 80% of rows. If columns are evenly distributed then it will be 0.8.
}
\description{
Get information about the columns in a dataset.
}
\examples{

dict(cars)

}
