% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixcat.R
\name{fixcat}
\alias{fixcat}
\title{Fix Categories}
\usage{
fixcat(x, minimum.cat.size = 15, other.name = "Other")
}
\arguments{
\item{x}{Data frame or tibble.}

\item{minimum.cat.size}{Minimum size for a category to not get grouped with other small categories.}

\item{other.name}{Value name for the group of small categories.}
}
\value{
Data frame with modified columns.
}
\description{
Groups small categories into a new bucket. Useful for preventing over-fitting on small categories.
}
\examples{
fixcat(iris)
}
