% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{genData_11}
\alias{genData_11}
\title{Generate data using support of size 11}
\usage{
genData_11(n, p, beta_true = NULL)
}
\arguments{
\item{n}{Integer value of sample size}

\item{p}{Integer number of variables}

\item{beta_true}{Input value of coefficients}
}
\value{
List of generated design matrix X, observation y, expected observation mu, 
and support supp
}
\description{
This function generates synthetic data, corresponding to model M5 in the paper.
}
