% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{genData_p}
\alias{genData_p}
\title{Generate data using support of size p}
\usage{
genData_p(n, p, beta_true = NULL)
}
\arguments{
\item{n}{Integer value of sample size}

\item{p}{Integer number of variables}

\item{beta_true}{Input value of coefficients}
}
\value{
List of generated design matrix X, observation y, expected observation mu, 
and support supp
}
\description{
This function generates synthetic data, corresponding to model M7 in the paper.
}
