% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genDesignMat.R
\name{genDesignMat}
\alias{genDesignMat}
\title{Return a design matrix}
\usage{
genDesignMat(n, p, r)
}
\arguments{
\item{n}{Integer value of sample size}

\item{p}{Integer number of variables}

\item{r}{Rate (0 to 1) of decay so the (i,j)th covariance is r^|i-j|}
}
\value{
Matrix of size n x p
}
\description{
This function generates a design matrix of given size,
whose rows are independent Gaussian with decaying covariance.
}
\examples{
X <- genDesignMat(150, 200, 0.5)
}
