% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCandidates.R
\name{getCandidates}
\alias{getCandidates}
\title{Get candidates from marginal ordering of variables}
\usage{
getCandidates(initCandidates, dat, weight_method = "ARM")
}
\arguments{
\item{initCandidates}{List of vectors each indicating an initial subset}

\item{dat}{List containing design Matrix X and observation Vector y}

\item{weight_method}{String indicating which method to use for weighting}
}
\value{
candidates List of vectors each representing a new candidate subset

w_var Vector of each variable's marginal importance

ind_var Vector of each variable's index in the original system 1...p
}
\description{
This function uses a weighting method to assign weights to 
each subset in a list of candidate subsets,
and then obtain marginal weights for variables to obtain 
a new candidate set with nested subsets.
}
