% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSupp.R
\name{getSupp}
\alias{getSupp}
\title{Get supports, i.e. the indices of selected variables, from a solution path}
\usage{
getSupp(Beta)
}
\arguments{
\item{Beta}{Matrix of size ('number of variables' + 1) x 'number of lambdas'}
}
\value{
List of 'number of lambdas' vectors, each representing a support 
(excluding intercept term)
}
\description{
This function uses a matrix Beta that represents a solution path to calculate
all the supports along solution path.
}
