% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{\%ni\%}
\alias{\%ni\%}
\title{Not In.}
\usage{
x \%ni\% y
}
\arguments{
\item{x}{Vector}

\item{y}{Vector}
}
\value{
Logical vector indicating which positions in x are not in y.
}
\description{
Easier alternative to ! x%in% y.
}
\examples{
#TODO
}
