% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash-files.R
\name{hash.files}
\alias{hash.files}
\title{Hash Files}
\usage{
hash.files(paths, hash.info = TRUE)
}
\arguments{
\item{paths}{List of file or folder locations.}

\item{hash.info}{Hash the file info ( FALSE hashes the contents ). This is faster but slightly less reliable.}
}
\value{
Hash string.
}
\description{
Create hash from a file or list of files. Useful for determining if files have changed during cache management.
}
\examples{
#TODO

}
