% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc.R
\name{model_performance}
\alias{model_performance}
\title{Model Performance}
\usage{
model_performance(data, model, transform_prob)
}
\arguments{
\item{data}{data-set to calculate performance}

\item{model}{model to calculate performance}

\item{transform_prob}{should probability be transformed before calculating performance
metrics?}
}
\description{
Calculates the model performance in a data-set
}

