% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelSelection.R
\name{modelSelection}
\alias{modelSelection}
\title{Model selection using different criteria}
\usage{
modelSelection(loglik, dim, n, criteria, penaltyBC = NULL)
}
\arguments{
\item{loglik}{Vector of (unnormalized) log-likelihood values evaluated for each model
evaluted at the maximum likelihood estimator}

\item{dim}{Vector of dimension (or number of free unknown parameters) of each model}

\item{n}{Integer of sample size}

\item{criteria}{Vector of strings ('AIC, BIC, BC') indicating criteria to use}

\item{penaltyBC}{Vector of non-default penalty values in using BC, default to NULL so that
only the suggested value n^(1/3) will be considered}
}
\value{
Vector (cri_opt) of indices of candidate models

Vector (cri_name) of the criteria used

Vector (PI) of parametricness index, default to one value that corresponds to 
the default BC penalty
}
\description{
This function uses specified criteria to select the optimal subset from a list of subsets, 
given design matrix X and observation y.
}
\examples{
res <- modelSelection(loglik = c(4.1, 5.2, 6.3), dim = c(1, 2, 3), n = 100, 
criteria='AIC, BIC, BC', penaltyBC=NULL)
}
