% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymodels.R
\name{my_randomForest}
\alias{my_randomForest}
\title{My Random Forest}
\usage{
my_randomForest(maxit = 10)
}
\arguments{
\item{maxit}{maximum number of iterations in random tune search}
}
\description{
This function implements my random forest tuning paremeters.
}

