% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nastrings.R
\docType{data}
\name{nastrings}
\alias{nastrings}
\title{Strings to consider as NA.}
\format{An object of class \code{character} of length 11.}
\usage{
nastrings
}
\description{
Common strings indicate data is not available. This dataset tracks these within my R package.
}
\examples{
# TODO
}
\keyword{datasets}
