% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc.R
\name{predict_model}
\alias{predict_model}
\title{Predict Model}
\usage{
predict_model(data, model, transform_prob = T)
}
\arguments{
\item{data}{dataset to predict}

\item{model}{model object}

\item{transform_prob}{should probability be transformed using the beta parameter?}
}
\description{
Predict Model
}

