% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRepExperiment.R
\name{runRepExperiment}
\alias{runRepExperiment}
\title{Run synthetic experiments for comparing the performance of high dimensional regression}
\usage{
runRepExperiment(genData, n, p, nRep, ratio, weight_method, fixedBeta)
}
\arguments{
\item{genData}{Function to generate synthetic training and testing data}

\item{n}{Integer value of sample size}

\item{p}{Integer number of variables}

\item{nRep}{Integer number (>1) of replications in the experiment}

\item{ratio}{Fraction (0 to 1) of the data used to generate solution paths}

\item{weight_method}{String ('simple' or 'ARM') of method to use in weighting subsets}

\item{fixedBeta}{Boolean indicator of whether to use fixed design in test}
}
\value{
Printout of means and standard errors of prediction loss, 
number of overfitted variables, and number of underfitted variables
}
\description{
This function generates synthetic data and run BCVO, LASSO, or others, 
and summarize its performance from several replications.
}
