% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sch.R
\name{sch}
\alias{sch}
\title{Search}
\usage{
sch(x, search.for, search.col = NULL, do.regex = FALSE,
  ignore.case = TRUE)
}
\arguments{
\item{x}{Data frame, tibble, or vector to search.}

\item{search.for}{String or value to search for.}

\item{search.col}{Column to search. If NULL, search all columns.}

\item{do.regex}{Use regular expression to search.}

\item{ignore.case}{Ignore case when searching.}
}
\value{
Search results.
}
\description{
Searches for a value and return rows/values that match.
}
\examples{
sch( iris, 'seto' )
}
