% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tonum.R
\name{tonum}
\alias{tonum}
\title{Convert to number.}
\usage{
tonum(x, na.strings = bc::na.strings, na.zero = FALSE,
  if.na = "return-unchanged", verbose = TRUE)
}
\arguments{
\item{x}{Vector to be covnerted to numeric.}

\item{na.strings}{Strings to consider NA.}

\item{na.zero}{Set NA values to 0.}

\item{if.na}{What to do if new NA values are crated during the conversion.}

\item{verbose}{Print messages (via cat) about what the function is doing.}
}
\value{
Result vector.
}
\description{
My flexible number conversion function.
}
\examples{
tonum( c( '1st', '2nd', '3rd' ) )
}
