% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc.R
\name{train_model}
\alias{train_model}
\title{Train Model}
\usage{
train_model(task, model, perc_perf = 0.2)
}
\arguments{
\item{task}{the task object}

\item{model}{the model object}

\item{perc_perf}{percentage of task size to be used to calculate performance}
}
\description{
This functions trains a model defined by a
mlr classifier. This is really just a wrapper mlr.
It carries the beta parameter that will be used in the
probability estimation later.
}
\details{
Its important to say that performance
}

