% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsel.R
\name{vsel}
\alias{vsel}
\title{Variable selector.}
\usage{
vsel(x, ycol, max.acceptable.pval = 0.1, verbose = FALSE)
}
\arguments{
\item{x}{full data frame including y/outcome/response column.}

\item{ycol}{Character value, name of the y/outcome/response column in the data.}

\item{max.acceptable.pval}{Maximum acceptable p-value for your model.}

\item{verbose}{Print information about status.}
}
\value{
Data with only selected variables.
}
\description{
Select variables to be used in a model. Should be called after normalization and splitting categorical to boolean.
}
\examples{
# TODO
}
