% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w.R
\name{w}
\alias{w}
\title{Write CSV Shorthand}
\usage{
w(x, file, row.names = FALSE, na = "", ...)
}
\arguments{
\item{x}{Data to write.}

\item{file}{Fielname, w will add .csv for you so you don't need to include it.}
}
\description{
Easier write function for quickly outputting data. Uses my preferred defaults.
}
\examples{
w( cars, 'output/cars' )
}
