% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash-files.R
\name{hash.files}
\alias{hash.files}
\title{Hash Files and Objects}
\usage{
hash.files(files)
}
\arguments{
\item{files}{List of file locations, or a list of objects to hash.}
}
\value{
Hash string.
}
\description{
Digest and hash a list of files. Useful for determining if files have changed during cache management.
}
\examples{
#TODO

}
