% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.any.R
\name{read.any}
\alias{read.any}
\title{Read Any}
\usage{
read.any(path, folder = NULL, sheet = 1)
}
\arguments{
\item{path}{Path to file within a folder, or full path.}

\item{folder}{Folder to look for path in, if applicable.}

\item{sheet}{If an excel file, the sheet to load.}
}
\value{
Tibble.
}
\description{
Flexible read function which also handles conversion to data types, and common data cleaning tasks.
}
\examples{
# TODO
}
