% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runfolder.R
\name{runfolder}
\alias{runfolder}
\title{Run the files in a folder.}
\usage{
runfolder(path, recursive = TRUE, verbose = TRUE)
}
\arguments{
\item{path}{Path to your files.}

\item{recursive}{Recursively runs child folders. See ?list.files.}

\item{verbose}{Print informative messages via cat.}
}
\description{
Helpful for looping through scripts.
}
\examples{
# TODO
}
