% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todate.R
\name{todate}
\alias{todate}
\title{Convert to Date.}
\usage{
todate(x, do.time = FALSE, do.excel = FALSE,
  na.vals = bc::na.strings, already.clean.vector = FALSE,
  verbose = TRUE, do.na = c("return-unchanged", "warning", "stop",
  "return-na"))
}
\arguments{
\item{x}{Vector to convert.}

\item{do.time}{NOT YET IMPLEMENTED Allow processing to time values.}

\item{do.excel}{NOT YET IMPLEMENTED Allow conversion of excel date integers.}

\item{already.clean.vector}{Sometimes in multi-step data operations clean.vector will have already been called.}

\item{verbose}{Print useful information via cat.}

\item{do.na}{What to do in case new NAs are created during conversion. If all NAs, the vector will be returned unchanged. Otherwise, this argument determines behavior: return-unchnaged returns the vector unchanged. warning returns new NAs with a warning. stop throws an error. return-na prints a message with cat which can be turned off with verbose = FALSE.}
}
\value{
Processed vector.
}
\description{
Convert a vector to date. By default, returns vector unchanged if new NAs would be created. 
You can send non-dates in loops (or otherwise) and they won't be affected. Can be useful for checking if something is a date.
}
\examples{
todate( c( '12/3/14', 'Hello'), do.na = 'return-na' )
todate( c( '12/3/14', 'Hello') )
}
